+++ crates/tinyfs/src/tests/memory.rs
@@ -1,4 +1,4 @@
-use crate::FS;
+use crate::fs::FS;
 use crate::error::Error;
 use crate::path_utils::normalize;
 use crate::path_utils::strip_root;
@@ -281,7 +281,7 @@ fn test_visit_glob_matching() {
     // Test case 2: Multiple match
     let paths: Vec<_> = root
         .visit("/a/file*.txt", |node, _| {
-            Ok(node.deref().read_file()?.to_vec())
+            Ok(node.derefX().read_file()?.to_vec())
         })
         .unwrap();
     assert_eq!(paths, vec![b"content1", b"content2"]);
@@ -289,7 +289,7 @@ fn test_visit_glob_matching() {
     // Test case 3: Multiple ** match
     let paths: Vec<_> = root
         .visit("/**/*.txt", |node, _| {
-            Ok(node.deref().read_file()?.to_vec())
+            Ok(node.derefX().read_file()?.to_vec())
         })
         .unwrap();
     assert_eq!(
@@ -306,14 +306,14 @@ fn test_visit_glob_matching() {
     // Test case 4: Single ** match
     let paths: Vec<_> = root
         .visit("/**/file4.txt", |node, _| {
-            Ok(node.deref().read_file()?.to_vec())
+            Ok(node.derefX().read_file()?.to_vec())
         })
         .unwrap();
     assert_eq!(paths, vec![b"content4"]);
 
     // Test case 5: Single ** match
     let paths: Vec<_> = root
-        .visit("/*/*.dat", |node, _| Ok(node.deref().read_file()?.to_vec()))
+        .visit("/*/*.dat", |node, _| Ok(node.derefX().read_file()?.to_vec()))
         .unwrap();
     assert_eq!(paths, vec![b"data"]);
 }
